minor fixes + switch merkle to merkle.old and merkle2 to merkle
