+++ services/statestorage/merkle.old/merkle_forest_test.go
@@ -1,77 +1,53 @@
-package merkle2
+package merkle
 
 import (
+	"encoding/base64"
 	"fmt"
 	"github.com/orbs-network/orbs-network-go/crypto/hash"
+	"github.com/orbs-network/orbs-network-go/test/builders"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
+	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/stretchr/testify/require"
 	"strings"
 	"testing"
 )
 
+//TODO - serialization based on spec (oded)
+//TODO - Radix 16 +/- parity
+
+//TODO - avoid hashing values of less than 32 bytes ?? Other optimizations (see ethereum)?
+//TODO - what hash functions should be used for values and what functions for node addresses?
+//TODO - should we include full values or just hashes (compare Ethereum)
+
 func updateStringEntries(f *Forest, baseHash primitives.MerkleSha256, keyValues ...string) primitives.MerkleSha256 {
 	if len(keyValues)%2 != 0 {
 		panic("expected key value pairs")
 	}
-	diffs := make(MerkleDiffs)
+	currentRoot := baseHash
 	for i := 0; i < len(keyValues); i = i + 2 {
-		diffs[keyValues[i]] = hash.CalcSha256([]byte(keyValues[i+1]))
-		// TODO fmt.Printf("new path %s value %x\n", keyValues[i], hash.CalcSha256([]byte(keyValues[i+1])))
+		currentRoot = f.updateSingleEntry(currentRoot, keyValues[i], hash.CalcSha256([]byte(keyValues[i+1])))
 	}
-
-	currentRoot, _ := f.Update(baseHash, diffs)
-
 	return currentRoot
 }
 
-func verifyProof(t *testing.T, f *Forest, root primitives.MerkleSha256, proof Proof, path string, value string, exists bool) {
-	verified, err := f.Verify(root, proof, path, hash.CalcSha256([]byte(value)))
+func verifyProof(t *testing.T, f *Forest, root primitives.MerkleSha256, proof Proof, contract string, key string, value string, exists bool) {
+	verified, err := f.Verify(root, proof, contract, key, value)
 	require.NoError(t, err, "proof verification failed")
 	require.Equal(t, exists, verified, "proof verification returned unexpected result")
 }
 
-func getProofRequireHeight(t *testing.T, f *Forest, root primitives.MerkleSha256, path string, expectedHeight int) Proof {
-	proof, err := f.GetProof(root, path)
+func getProofRequireHeight(t *testing.T, f *Forest, root primitives.MerkleSha256, contract string, key string, expectedHeight int) Proof {
+	proof, err := f.GetProof(root, contract, key)
 	require.NoError(t, err, "failed with error: %s", err)
-	require.Equal(t, expectedHeight, len(proof), "unexpected proof length")
+	require.Len(t, proof, expectedHeight, "unexpected proof length")
 	return proof
-}
-
-func TestRootManagement(t *testing.T) {
-	f, _ := NewForest()
-
-	require.Len(t, f.roots, 1, "new forest should have 1 root")
-	emptyNode := createEmptyNode()
-	foundRoot := f.findRoot(emptyNode.hash)
-	require.Equal(t, emptyNode, foundRoot, "proof verification returned unexpected result")
-
-	node1 := createNode("hi", hash.CalcSha256([]byte("bye")), true)
-	node1.hash = node1.serialize().hash()
-	node2 := createNode("bye", hash.CalcSha256([]byte("d")), true)
-	node2.hash = node2.serialize().hash()
-
-	f.appendRoot(node1)
-	f.appendRoot(node2)
-	require.Len(t, f.roots, 3, "mismatch length")
-
-	node1hash := createNode("hi", hash.CalcSha256([]byte("bye")), true).serialize().hash()
-	foundRoot = f.findRoot(node1hash)
-	require.Equal(t, node1, foundRoot, "should be same node")
-
-	node2hash := createNode("bye", hash.CalcSha256([]byte("d")), true).serialize().hash()
-	f.Forget(node2hash)
-	require.Len(t, f.roots, 2, "mismatch length after forget")
-	foundRoot = f.findRoot(node1.hash)
-	require.Equal(t, node1, foundRoot, "should be same node")
-	foundRoot = f.findRoot(node2.hash)
-	require.Nil(t, foundRoot, "node should not exist")
 }
 
 func TestAddSingleEntryToEmptyTree(t *testing.T) {
 	f, root := NewForest()
 	root = updateStringEntries(f, root, "bar", "baz")
 
-	getProofRequireHeight(t, f, root, "bar", 1)
+	getProofRequireHeight(t, f, root, "", "bar", 1)
 }
 
 func TestRootChangeAfterStateChange(t *testing.T) {
@@ -81,7 +57,6 @@ func TestRootChangeAfterStateChange(t *testing.T) {
 	root2 := updateStringEntries(f, root1, "first", "val1")
 
 	require.NotEqual(t, root1, root2, "root hash did not change after state change")
-	// TODO add verify here ?
 }
 
 func TestRevertingStateChangeRevertsMerkleRoot(t *testing.T) {
@@ -94,55 +69,62 @@ func TestRevertingStateChangeRevertsMerkleRoot(t *testing.T) {
 	require.Equal(t, root1, root3, "root hash did not revert back after resetting state")
 }
 
-
 func TestValidProofForMissingKey(t *testing.T) {
 	f, root := NewForest()
 	key := "imNotHere"
-	proof := getProofRequireHeight(t, f, root, key, 1)
-	verifyProof(t, f, root, proof, key, "", true)
-	verifyProof(t, f, root, proof, key, "non-zero", false)
+	contract := "foo"
+	proof := getProofRequireHeight(t, f, root, contract, key, 1)
+	verifyProof(t, f, root, proof, contract, key, "", true)
+	verifyProof(t, f, root, proof, contract, key, "non-zero", false)
 
 }
 
 func TestUpdateTrieFailsForMissingBaseNode(t *testing.T) {
-	f, _ := NewForest()
-	badroot := primitives.MerkleSha256(hash.CalcSha256([]byte("justanytext")))
+	f, root := NewForest()
 
-	root := updateStringEntries(f, badroot, "first", "val")
-
-	require.Nil(t, root, "did not receive an empty response when using a corrupt merkle root")
+	root[3] = root[3] + 1
+	csd := builders.ContractStateDiff().WithContractName("foo").WithStringRecord("bar1", "baz").Build()
+	_, err := f.Update(root, []*protocol.ContractStateDiff{csd})
+	require.Error(t, err, "did not receive an error when using a corrupt merkle.old root")
 }
 
 func TestProofValidationAfterBatchStateUpdate(t *testing.T) {
 	f, root := NewForest()
 
-	root1 := updateStringEntries(f, root, "bar1", "baz", "shared", "quux1")
+	csd1 := builders.ContractStateDiff().WithContractName("foo").
+		WithStringRecord("bar1", "baz").WithStringRecord("shared", "quux1").Build()
+	iterator1 := csd1.StateDiffsIterator()
+	bar1 := iterator1.NextStateDiffs()
+	shared1 := iterator1.NextStateDiffs()
 
-	proof := getProofRequireHeight(t, f, root1, "bar1", 2)
-	verifyProof(t, f, root1, proof, "bar1", "baz", true)
-	proof = getProofRequireHeight(t, f, root1, "bar2", 2)
-	verifyProof(t, f, root1, proof, "bar2", "qux", false)
+	csd2 := builders.ContractStateDiff().WithContractName("foo").
+		WithStringRecord("bar2", "qux").WithStringRecord("shared", "quux2").Build()
+	iterator2 := csd2.StateDiffsIterator()
+	bar2 := iterator2.NextStateDiffs()
+	shared2 := iterator2.NextStateDiffs()
 
-	root2 := updateStringEntries(f, root1, "bar2", "qux", "shared", "quux2")
-	require.NotEqual(t, root2, root1, "roots should be different")
+	root1, _ := f.Update(root, []*protocol.ContractStateDiff{csd1})
+	root2, _ := f.Update(root1, []*protocol.ContractStateDiff{csd2})
 
-	// retest that first insert proof still valid
-	proof = getProofRequireHeight(t, f, root1, "bar1", 2)
-	verifyProof(t, f, root1, proof, "bar1", "baz", true)
-	proof = getProofRequireHeight(t, f, root1, "bar2", 2)
-	verifyProof(t, f, root1, proof, "bar2", "qux", false)
-	proof = getProofRequireHeight(t, f, root1, "shared", 2)
-	verifyProof(t, f, root1, proof, "shared", "quux1", true)
-	verifyProof(t, f, root1, proof, "shared", "quux2", false)
+	proof := getProofRequireHeight(t, f, root1, "foo", bar1.StringKey(), 2)
+	verifyProof(t, f, root1, proof, "foo", bar1.StringKey(), bar1.StringValue(), true)
 
-	// after second insert proofs
-	proof = getProofRequireHeight(t, f, root2, "bar2", 3)
-	verifyProof(t, f, root2, proof, "bar2", "qux", true)
-	proof = getProofRequireHeight(t, f, root2, "bar1", 3)
-	verifyProof(t, f, root2, proof, "bar1", "baz", true)
-	proof = getProofRequireHeight(t, f, root2, "shared", 2)
-	verifyProof(t, f, root2, proof, "shared", "quux2", true)
-	verifyProof(t, f, root2, proof, "shared", "quux1", false)
+	proof = getProofRequireHeight(t, f, root1, "foo", bar2.StringKey(), 2)
+	verifyProof(t, f, root1, proof, "foo", bar2.StringKey(), bar2.StringValue(), false)
+
+	proof = getProofRequireHeight(t, f, root2, "foo", bar2.StringKey(), 3)
+	verifyProof(t, f, root2, proof, "foo", bar2.StringKey(), bar2.StringValue(), true)
+
+	proof = getProofRequireHeight(t, f, root2, "foo", bar1.StringKey(), 3)
+	verifyProof(t, f, root2, proof, "foo", bar1.StringKey(), bar1.StringValue(), true)
+
+	proof = getProofRequireHeight(t, f, root1, "foo", shared1.StringKey(), 2)
+	verifyProof(t, f, root1, proof, "foo", shared1.StringKey(), shared1.StringValue(), true)
+	verifyProof(t, f, root1, proof, "foo", shared1.StringKey(), shared2.StringValue(), false)
+
+	proof = getProofRequireHeight(t, f, root2, "foo", shared2.StringKey(), 2)
+	verifyProof(t, f, root2, proof, "foo", shared2.StringKey(), shared2.StringValue(), true)
+	verifyProof(t, f, root2, proof, "foo", shared2.StringKey(), shared1.StringValue(), false)
 }
 
 func TestProofValidationForTwoRevisionsOfSameKey(t *testing.T) {
@@ -150,42 +132,26 @@ func TestProofValidationForTwoRevisionsOfSameKey(t *testing.T) {
 	root1 := updateStringEntries(f, root, "bar1", "baz1")
 	root2 := updateStringEntries(f, root1, "bar1", "baz2")
 
-	proof := getProofRequireHeight(t, f, root1, "bar1", 1)
-	verifyProof(t, f, root1, proof, "bar1", "baz1", true)
+	proof := getProofRequireHeight(t, f, root1, "", "bar1", 1)
+	verifyProof(t, f, root1, proof, "", "bar1", "baz1", true)
 
-	proof = getProofRequireHeight(t, f, root2, "bar1", 1)
-	verifyProof(t, f, root2, proof, "bar1", "baz2", true)
+	proof = getProofRequireHeight(t, f, root2, "", "bar1", 1)
+	verifyProof(t, f, root2, proof, "", "bar1", "baz2", true)
 }
 
-func TestExtendingLeafNodeWithNoBranchesOneAtATime(t *testing.T) {
-	f, root := NewForest()
-	root1 := updateStringEntries(f, root, "ba", "zoo")
-	root2 := updateStringEntries(f, root1, "bar", "baz")
-	root3 := updateStringEntries(f, root2, "baron", "Hello")
-
-	getProofRequireHeight(t, f, root3, "baron", 3)
-}
-
-func TestExtendingLeafNodeWithNoBranches(t *testing.T) {
+func TestExtendingLeafNodeWithNoBranchesAndNoValue(t *testing.T) {
 	f, root := NewForest()
 	root = updateStringEntries(f, root, "ba", "zoo", "bar", "baz", "baron", "Hello")
 
-	getProofRequireHeight(t, f, root, "baron", 3)
-}
-
-func TestExtendingLeafNodeWithNoBranchesInWrongOrder(t *testing.T) {
-	f, root := NewForest()
-	root = updateStringEntries(f, root, "baron", "Hello", "bar", "baz", "ba", "zoo")
-
-	getProofRequireHeight(t, f, root, "baron", 3)
+	getProofRequireHeight(t, f, root, "", "baron", 3)
 }
 
 func TestExtendingKeyPathByOneChar(t *testing.T) {
 	f, root := NewForest()
 	root = updateStringEntries(f, root, "bar", "baz", "bar1", "qux")
 
-	proof := getProofRequireHeight(t, f, root, "bar1", 2)
-	verifyProof(t, f, root, proof, "bar1", "qux", true)
+	proof := getProofRequireHeight(t, f, root, "", "bar1", 2)
+	verifyProof(t, f, root, proof, "", "bar1", "qux", true)
 }
 
 func TestExtendingKeyPathBySeveralChars(t *testing.T) {
@@ -193,33 +159,33 @@ func TestExtendingKeyPathBySeveralChars(t *testing.T) {
 
 	root1 := updateStringEntries(f, root, "bar", "baz", "bar12", "qux", "bar123456789", "quux")
 
-	proof := getProofRequireHeight(t, f, root1, "bar123456789", 3)
-	verifyProof(t, f, root1, proof, "bar123456789", "quux", true)
+	proof := getProofRequireHeight(t, f, root1, "", "bar123456789", 3)
+	verifyProof(t, f, root1, proof, "", "bar123456789", "quux", true)
 }
 
 func TestAddSiblingNode(t *testing.T) {
 	f, root := NewForest()
 	root1 := updateStringEntries(f, root, "bar", "baz", "bar1", "qux", "bar2", "quux")
 
-	proof := getProofRequireHeight(t, f, root1, "bar2", 2)
-	verifyProof(t, f, root1, proof, "bar2", "quux", true)
+	proof := getProofRequireHeight(t, f, root1, "", "bar2", 2)
+	verifyProof(t, f, root1, proof, "", "bar2", "quux", true)
 }
 
 func TestAddPathToCauseBranchingAlongExistingPath(t *testing.T) {
 	f, root := NewForest()
 	root1 := updateStringEntries(f, root, "bar", "baz", "bar1", "qux", "bad", "quux")
 
-	proof := getProofRequireHeight(t, f, root1, "bad", 2)
-	verifyProof(t, f, root1, proof, "bad", "quux", true)
+	proof := getProofRequireHeight(t, f, root1, "", "bad", 2)
+	verifyProof(t, f, root1, proof, "", "bad", "quux", true)
 }
 
 func TestReplaceExistingValueBelowDivergingPaths(t *testing.T) {
 	f, root := NewForest()
 	root1 := updateStringEntries(f, root, "bar", "baz", "bar1", "qux", "bad", "quux", "bar1", "zoo")
 
-	proof := getProofRequireHeight(t, f, root1, "bar1", 3)
-	verifyProof(t, f, root1, proof, "bar1", "zoo", true)
-	verifyProof(t, f, root1, proof, "bar1", "qux", false)
+	proof := getProofRequireHeight(t, f, root1, "", "bar1", 3)
+	verifyProof(t, f, root1, proof, "", "bar1", "zoo", true)
+	verifyProof(t, f, root1, proof, "", "bar1", "qux", false)
 }
 
 func TestAddPathToCauseNewLeafAlongExistingPath(t *testing.T) {
@@ -227,22 +193,22 @@ func TestAddPathToCauseNewLeafAlongExistingPath(t *testing.T) {
 
 	root1 := updateStringEntries(f, root, "baron", "Hirsch", "bar", "Hello")
 
-	proof := getProofRequireHeight(t, f, root1, "bar", 1)
-	verifyProof(t, f, root1, proof, "bar", "Hello", true)
+	proof := getProofRequireHeight(t, f, root1, "", "bar", 1)
+	verifyProof(t, f, root1, proof, "", "bar", "Hello", true)
 
-	proof = getProofRequireHeight(t, f, root1, "baron", 2)
-	verifyProof(t, f, root1, proof, "baron", "Hirsch", true)
+	proof = getProofRequireHeight(t, f, root1, "", "baron", 2)
+	verifyProof(t, f, root1, proof, "", "baron", "Hirsch", true)
 }
 
 func TestRemoveValue_SingleExistingNode(t *testing.T) {
 	f, root := NewForest()
 
 	root1 := updateStringEntries(f, root, "aKey", "aValue")
-	root2 := updateStringEntries(f, root1, "aKey", "")
+	root2 := updateStringEntries(f, root, "aKey", "")
 
-	getProofRequireHeight(t, f, root, "aKey", 1)
-	getProofRequireHeight(t, f, root1, "aKey", 1)
-	getProofRequireHeight(t, f, root2, "aKey", 1)
+	getProofRequireHeight(t, f, root, "", "aKey", 1)
+	getProofRequireHeight(t, f, root1, "", "aKey", 1)
+	getProofRequireHeight(t, f, root2, "", "aKey", 1)
 	require.EqualValues(t, root, root2, "for identical states hash must be identical")
 	require.NotEqual(t, root1, root2, "for different states hash must be different")
 }
@@ -254,9 +220,9 @@ func TestRemoveValue_RemoveSingleChildLeaf(t *testing.T) {
 	root2 := updateStringEntries(f, root1, "prefixSuffix", "2")
 	root3 := updateStringEntries(f, root2, "prefixSuffix", "")
 
-	getProofRequireHeight(t, f, root1, "prefixSuffix", 1)
-	getProofRequireHeight(t, f, root2, "prefixSuffix", 2)
-	getProofRequireHeight(t, f, root3, "prefixSuffix", 1)
+	getProofRequireHeight(t, f, root1, "", "prefixSuffix", 1)
+	getProofRequireHeight(t, f, root2, "", "prefixSuffix", 2)
+	getProofRequireHeight(t, f, root3, "", "prefixSuffix", 1)
 	require.EqualValues(t, root1, root3, "root hash should be identical")
 }
 
@@ -265,7 +231,7 @@ func TestRemoveValue_ParentWithSingleChild(t *testing.T) {
 
 	root1 := updateStringEntries(f, root, "no", "1", "noam", "1", "no", "")
 
-	p := getProofRequireHeight(t, f, root1, "noam", 1)
+	p := getProofRequireHeight(t, f, root1, "", "noam", 1)
 	require.EqualValues(t, "noam", p[0].path, "full tree proof for and does not end with expected node path")
 }
 
@@ -275,11 +241,11 @@ func TestRemoveValue_NonBranchingNonLeaf1(t *testing.T) {
 	fullTree := updateStringEntries(f, root, "a", "1", "and", "2", "android", "3")
 	afterRemove := updateStringEntries(f, fullTree, "and", "")
 
-	p1 := getProofRequireHeight(t, f, fullTree, "and", 2)
-	p2 := getProofRequireHeight(t, f, afterRemove, "and", 2)
+	p1 := getProofRequireHeight(t, f, fullTree, "", "and", 2)
+	p2 := getProofRequireHeight(t, f, afterRemove, "", "and", 2)
 
-	getProofRequireHeight(t, f, fullTree, "android", 3)
-	getProofRequireHeight(t, f, afterRemove, "android", 2)
+	getProofRequireHeight(t, f, fullTree, "", "android", 3)
+	getProofRequireHeight(t, f, afterRemove, "", "android", 2)
 
 	require.EqualValues(t, "d", p1[1].path, "full tree proof for and does not end with expected node path")
 	require.EqualValues(t, "droid", p2[1].path, "full tree proof for and does not end with expected node path")
@@ -291,11 +257,11 @@ func TestRemoveValue_NonBranchingNonLeaf2(t *testing.T) {
 	fullTree := updateStringEntries(f, root, "an", "1", "and", "2", "android", "3")
 	afterRemove := updateStringEntries(f, fullTree, "and", "")
 
-	p1 := getProofRequireHeight(t, f, fullTree, "and", 2)
-	p2 := getProofRequireHeight(t, f, afterRemove, "and", 2)
+	p1 := getProofRequireHeight(t, f, fullTree, "", "and", 2)
+	p2 := getProofRequireHeight(t, f, afterRemove, "", "and", 2)
 
-	getProofRequireHeight(t, f, fullTree, "android", 3)
-	getProofRequireHeight(t, f, afterRemove, "android", 2)
+	getProofRequireHeight(t, f, fullTree, "", "android", 3)
+	getProofRequireHeight(t, f, afterRemove, "", "android", 2)
 
 	require.EqualValues(t, "", p1[1].path, "full tree proof for and does not end with expected node path")
 	require.EqualValues(t, "roid", p2[1].path, "full tree proof for and does not end with expected node path")
@@ -307,11 +273,11 @@ func TestRemoveValue_BranchingNonLeaf_NodeStructureUnchanged(t *testing.T) {
 	fullTree := updateStringEntries(f, root, "and", "1", "andalusian", "1", "android", "1")
 	afterRemove := updateStringEntries(f, fullTree, "and", "")
 
-	p1 := getProofRequireHeight(t, f, afterRemove, "andalusian", 2)
-	p2 := getProofRequireHeight(t, f, afterRemove, "android", 2)
+	p1 := getProofRequireHeight(t, f, afterRemove, "", "andalusian", 2)
+	p2 := getProofRequireHeight(t, f, afterRemove, "", "android", 2)
 
-	getProofRequireHeight(t, f, fullTree, "android", 2)
-	getProofRequireHeight(t, f, afterRemove, "android", 2)
+	getProofRequireHeight(t, f, fullTree, "", "android", 2)
+	getProofRequireHeight(t, f, afterRemove, "", "android", 2)
 
 	require.EqualValues(t, "lusian", p1[1].path, "full tree proof for and does not end with expected node path")
 	require.EqualValues(t, "oid", p2[1].path, "full tree proof for and does not end with expected node path")
@@ -323,11 +289,11 @@ func TestRemoveValue_BranchingNonLeaf_CollapseBranch(t *testing.T) {
 	root1 := updateStringEntries(f, root, "no", "7", "noam", "8", "noan", "9")
 	root2 := updateStringEntries(f, root1, "no", "")
 
-	p0 := getProofRequireHeight(t, f, root1, "noam", 3)
+	p0 := getProofRequireHeight(t, f, root1, "", "noam", 3)
 	require.EqualValues(t, "no", p0[0].path, "unexpected proof structure")
 
-	p := getProofRequireHeight(t, f, root2, "noam", 2)
-	// TODO require.EqualValues(t, false, p[0].hasValue(), "unexpected proof structure")
+	p := getProofRequireHeight(t, f, root2, "", "noam", 2)
+	require.EqualValues(t, false, p[0].hasValue(), "unexpected proof structure")
 	require.EqualValues(t, "noa", p[0].path, "unexpected proof structure")
 }
 
@@ -337,8 +303,8 @@ func TestRemoveValue_OneOfTwoChildren(t *testing.T) {
 	root1 := updateStringEntries(f, root, "noa", "1", "noam", "1", "noan", "1")
 	root2 := updateStringEntries(f, root1, "noan", "")
 
-	p := getProofRequireHeight(t, f, root2, "noam", 2)
-	getProofRequireHeight(t, f, root2, "noan", 1)
+	p := getProofRequireHeight(t, f, root2, "", "noam", 2)
+	getProofRequireHeight(t, f, root2, "", "noan", 1)
 	require.EqualValues(t, "noa", p[0].path, "full tree proof for and does not end with expected node path")
 }
 
@@ -347,8 +313,8 @@ func TestRemoveValue_OneOfTwoChildrenCollapsingParent(t *testing.T) {
 
 	root1 := updateStringEntries(f, root, "noam", "8", "noan", "9", "noan", "")
 
-	p := getProofRequireHeight(t, f, root1, "noam", 1)
-	getProofRequireHeight(t, f, root1, "noan", 1)
+	p := getProofRequireHeight(t, f, root1, "", "noam", 1)
+	getProofRequireHeight(t, f, root1, "", "noan", 1)
 	require.EqualValues(t, "noam", p[0].path, "unexpected proof structure")
 }
 
@@ -376,12 +342,12 @@ func TestOrderOfAdditionsDoesNotMatter(t *testing.T) {
 	f1, initRoot1 := NewForest()
 	root1 := updateStringEntries(f1, initRoot1, keyValue[var1[0]], keyValue[var1[0]+1], keyValue[var1[1]], keyValue[var1[1]+1],
 		keyValue[var1[2]], keyValue[var1[2]+1], keyValue[var1[3]], keyValue[var1[3]+1], keyValue[var1[4]], keyValue[var1[4]+1])
-	proof1, _ := f1.GetProof(root1, "bar1234")
+	proof1, _ := f1.GetProof(root1, "", "bar1234")
 
 	f2, initRoot2 := NewForest()
 	root2 := updateStringEntries(f2, initRoot2, keyValue[var2[0]], keyValue[var2[0]+1], keyValue[var2[1]], keyValue[var2[1]+1],
 		keyValue[var2[2]], keyValue[var2[2]+1], keyValue[var2[3]], keyValue[var2[3]+1], keyValue[var2[4]], keyValue[var2[4]+1])
-	proof2, _ := f2.GetProof(root2, "bar1234")
+	proof2, _ := f2.GetProof(root2, "", "bar1234")
 
 	require.Equal(t, root1, root2, "unexpected different root hash")
 	require.Equal(t, len(proof1), len(proof2), "unexpected different tree depth / proof lengths")
@@ -390,7 +356,7 @@ func TestOrderOfAdditionsDoesNotMatter(t *testing.T) {
 	f3, initRoot3 := NewForest()
 	root3 := updateStringEntries(f3, initRoot3, keyValue[var3[0]], keyValue[var3[0]+1], keyValue[var3[1]], keyValue[var3[1]+1],
 		keyValue[var3[2]], keyValue[var3[2]+1], keyValue[var3[3]], keyValue[var3[3]+1], keyValue[var3[4]], keyValue[var3[4]+1])
-	proof3, _ := f3.GetProof(root3, "bar1234")
+	proof3, _ := f3.GetProof(root3, "", "bar1234")
 
 	require.Equal(t, root2, root3, "unexpected different root hash")
 	require.Equal(t, len(proof2), len(proof3), "unexpected different tree depth / proof lengths")
@@ -402,16 +368,20 @@ func TestAddConvegingPathsWithExactValues(t *testing.T) {
 	root1 := updateStringEntries(f, root, "abdbda", "1", "abdcda", "1", "acdbda", "1", "acdcda", "1")
 	root2 := updateStringEntries(f, root1, "abdcda", "2")
 
-	proof1, _ := f.GetProof(root2, "abdbda")
-	proof2, _ := f.GetProof(root2, "abdcda")
-	proof3, _ := f.GetProof(root2, "acdbda")
-	proof4, _ := f.GetProof(root2, "acdcda")
+	proof1, _ := f.GetProof(root2, "", "abdbda")
+	proof2, _ := f.GetProof(root2, "", "abdcda")
+	proof3, _ := f.GetProof(root2, "", "acdbda")
+	proof4, _ := f.GetProof(root2, "", "acdcda")
 
-	verifyProof(t, f, root2, proof1, "abdbda", "1", true)
-	verifyProof(t, f, root2, proof2, "abdcda", "2", true)
-	verifyProof(t, f, root2, proof3, "acdbda", "1", true)
-	verifyProof(t, f, root2, proof4, "acdcda", "1", true)
+	proof3.dump()
+	proof4.dump()
+
+	verifyProof(t, f, root2, proof1, "", "abdbda", "1", true)
+	verifyProof(t, f, root2, proof2, "", "abdcda", "2", true)
+	verifyProof(t, f, root2, proof3, "", "acdbda", "1", true)
+	verifyProof(t, f, root2, proof4, "", "acdcda", "1", true)
 }
+
 
 
 // =================
@@ -420,13 +390,24 @@ func TestAddConvegingPathsWithExactValues(t *testing.T) {
 
 func (f *Forest) dump(t *testing.T) {
 	t.Logf("---------------- TRIE BEGIN ------------------")
-	for _, root := range f.roots {
-		root.printNode(" Ω", 0, f, t)
+	childNodes := make(map[string]*Node, len(f.nodes))
+	for _, n := range f.nodes {
+		for _, childHash := range n.branches {
+			if len(childHash) != 0 {
+				childNodes[childHash.KeyForMap()] = f.nodes[childHash.KeyForMap()]
+			}
+		}
+	}
+
+	for nodeHash := range f.nodes {
+		if _, isChild := childNodes[nodeHash]; !isChild {
+			f.nodes[nodeHash].printNode(" Ω", 0, f, t)
+		}
 	}
 	t.Logf("---------------- TRIE END --------------------")
 }
 
-func (n *node) printNode(label string, depth int, trie *Forest, t *testing.T) {
+func (n *Node) printNode(label string, depth int, trie *Forest, t *testing.T) {
 	prefix := strings.Repeat(" ", depth)
 	leafText := ""
 	if n.hasValue() {
@@ -435,8 +416,24 @@ func (n *node) printNode(label string, depth int, trie *Forest, t *testing.T) {
 	pathString := fmt.Sprintf("%s%s)%s", prefix, label, n.path)
 	t.Logf("%s%s\n", pathString, leafText)
 	for l, v := range n.branches {
-		if v != nil {
-			v.printNode(string([]byte{byte(l)}), depth+len(pathString)-1, trie, t)
+		if len(v) != 0 {
+			trie.nodes[v.KeyForMap()].printNode(string([]byte{byte(l)}), depth+len(pathString)-1, trie, t)
 		}
+	}
+}
+
+func (p *Proof) dump() {
+	fmt.Println("---------------- PROOF BEGIN ------------------")
+	for _, n := range *p {
+		hash2 := n.hash()
+		fmt.Printf("%s\n%+v\n", base64.StdEncoding.EncodeToString(hash2[:]), n)
+	}
+	fmt.Println("---------------- PROOF END --------------------")
+}
+
+// TODO - this just checks there are no data integrity in our driver integrity
+func (f *Forest) testForestIntegrity(t *testing.T) {
+	for h, n := range f.nodes {
+		require.Equal(t, n.hash().KeyForMap(), h, "node key is not true hash code")
 	}
 }
