+++ services/statestorage/merkle/_merkle_contract_test.go
@@ -1,9 +1,9 @@
-package merkle2
+package merkle
 
 import (
 	"fmt"
 	"github.com/orbs-network/orbs-network-go/crypto/hash"
-	"github.com/orbs-network/orbs-network-go/services/statestorage/merkle"
+	"github.com/orbs-network/orbs-network-go/services/statestorage/merkle.old"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
 	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/stretchr/testify/require"
