+++ services/statestorage/merkle/merkle_forest.go
@@ -4,231 +4,155 @@ import (
 	"fmt"
 	"github.com/orbs-network/orbs-network-go/crypto/hash"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
-	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/pkg/errors"
 	"strings"
+	"sync"
 )
-
-type Proof []*Node
 
 const trieRadix = 256 // base of the merkle trie. TODO change to 16
 
-func getZeroValueHash() primitives.Sha256 {
+func GetZeroValueHash() primitives.Sha256 {
 	return hash.CalcSha256([]byte{})
 }
 
-var zeroValueHash = getZeroValueHash()
+var zeroValueHash = GetZeroValueHash()
 
-type Node struct {
+type Proof []*ProofNode
+
+// TODO replace proofNode with membuf/proto
+type ProofNode struct {
 	path     string // TODO replace with []byte + parity bool when moving to trieRadix = 16
 	value    primitives.Sha256
 	branches [trieRadix]primitives.MerkleSha256
 }
 
-func createNode(path string, valueHash primitives.Sha256) *Node {
-	return &Node{
-		path:     path,
-		value:    valueHash,
-		branches: [trieRadix]primitives.MerkleSha256{},
-	}
-}
-func (n *Node) hasValue() bool {
-	return !zeroValueHash.Equal(n.value)
-}
-func (n *Node) hash() primitives.MerkleSha256 {
-	// TODO replace this with membuffers serialization. Sprintf should not reach production
-	serializedNode := fmt.Sprintf("%+v", n)
+func (pn *ProofNode) hash() primitives.MerkleSha256 {
+	serializedNode := fmt.Sprintf("%+v", pn)
 	return primitives.MerkleSha256(hash.CalcSha256([]byte(serializedNode)))
 }
-func (n *Node) clone() *Node {
-	newBranches := [trieRadix]primitives.MerkleSha256{}
-	copy(newBranches[:], n.branches[:])
-	result := &Node{
+
+type node struct {
+	path     string // TODO replace with []byte + parity bool when moving to trieRadix = 16
+	value    primitives.Sha256
+	hash     primitives.MerkleSha256
+	branches [trieRadix]*node
+	isLeaf   bool
+}
+
+func createNode(path string, valueHash primitives.Sha256, isLeaf bool) *node {
+	return &node{
+		path:     path,
+		value:    valueHash,
+		branches: [trieRadix]*node{},
+		isLeaf:   isLeaf,
+		hash:     primitives.MerkleSha256{},
+	}
+}
+
+func createEmptyNode() *node {
+	tmp := createNode("", zeroValueHash, true)
+	tmp.hash = tmp.serialize().hash()
+	return tmp
+}
+
+func (n *node) hasValue() bool {
+	return !zeroValueHash.Equal(n.value)
+}
+
+func (n *node) serialize() *ProofNode {
+	sn := &ProofNode{
+		path:     n.path,
+		value:    n.value,
+		branches: [trieRadix]primitives.MerkleSha256{},
+	}
+	if !n.isLeaf {
+		for k, v := range n.branches {
+			if v != nil {
+				sn.branches[k] = v.hash
+			}
+		}
+	}
+	return sn
+}
+
+func (n *node) clone() *node {
+	newBranches := [trieRadix]*node{}
+	if !n.isLeaf {
+		copy(newBranches[:], n.branches[:])
+	}
+	result := &node{
 		path:     n.path,
 		value:    n.value, // TODO - copy?
 		branches: newBranches,
+		isLeaf:   n.isLeaf,
 	}
-	//for k, v := range n.branches {
-	//	result.branches[k] = v // TODO - copy?
-	//}
 	return result
 }
 
-func (n *Node) hasChildren() bool {
-	for _, v := range n.branches {
-		if v != nil {
-			return true
-		}
-	}
-	return false
+type dirtyNodes map[*node]map[byte]bool
+
+func (dn dirtyNodes) init(node *node) {
+	dn[node] = make(map[byte]bool)
 }
 
-func (n *Node) getSingleChildSelector() *byte {
-	var singleChildSelector *byte
-	for i, v := range n.branches {
-		if v != nil {
-			if singleChildSelector != nil {
-				return nil
-			}
-			ib := byte(i)
-			singleChildSelector = &ib
-		}
+func (dn dirtyNodes) set(node *node, arc byte) {
+	if _, exits := dn[node]; !exits {
+		dn.init(node)
 	}
-	return singleChildSelector
+	dn[node][arc] = true
 }
+
+type MerkleDiffs map[string]primitives.Sha256
 
 type Forest struct {
-	nodes map[string]*Node
+	mutex sync.Mutex
+	roots []*node
 }
 
-// return the merkle trie & the trie root hash for the empty default trie
 func NewForest() (*Forest, primitives.MerkleSha256) {
-	var emptyNode = createNode("", zeroValueHash)
-	var emptyNodeHash = emptyNode.hash()
-	return &Forest{nodes: map[string]*Node{emptyNodeHash.KeyForMap(): emptyNode}}, emptyNodeHash
+	var emptyNode = createEmptyNode()
+	return &Forest{sync.Mutex{}, []*node{emptyNode}}, emptyNode.hash
 }
 
-func (f *Forest) connectChildToParentAndSaveChild(childNode, parentNode *Node, selector byte) {
-	childHash := childNode.hash()
-	parentNode.branches[selector] = childHash
-	f.nodes[childHash.KeyForMap()] = childNode
-}
+func (f *Forest) findRoot(rootHash primitives.MerkleSha256) *node {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-func (f *Forest) updateSingleEntry(baseHash primitives.MerkleSha256, path string, valueHash primitives.Sha256) primitives.MerkleSha256 {
-	// TODO fmt.Printf("old path %s value %x\n", path, valueHash)
-	baseNode := f.nodes[baseHash.KeyForMap()]
-	var newRoot *Node
-	if valueHash.Equal(zeroValueHash) {
-		newRoot = f.remove(baseNode, path)
-	} else {
-		newRoot = f.add(baseNode, path, valueHash)
-	}
-	sha256 := newRoot.hash()
-	f.nodes[sha256.KeyForMap()] = newRoot
-	return sha256
-}
-
-func (f *Forest) squash(n *Node) *Node {
-	if !n.hasValue() && !n.hasChildren() && n.path != "" { // no branch and no value - reset path
-		clone := n.clone()
-		clone.path = ""
-		return clone
-	}
-	if singleChildSelector := n.getSingleChildSelector(); singleChildSelector != nil && !n.hasValue() { // merge with single child
-		clone := f.nodes[n.branches[*singleChildSelector].KeyForMap()].clone()
-		clone.path = n.path + string([]byte{*singleChildSelector}) + clone.path
-		return clone
-	}
-	return n
-}
-
-func (f *Forest) remove(currentNode *Node, path string) *Node {
-	if currentNode.path == path { // reached node with value that is being zeroed
-		clone := currentNode.clone()
-		clone.value = zeroValueHash
-		clone = f.squash(clone)
-		return clone
-	}
-
-	if strings.HasPrefix(path, currentNode.path) {
-		clone := currentNode.clone()
-		branchSelector := path[len(currentNode.path)]
-		branchHash := clone.branches[branchSelector]
-		if branchHash != nil {
-			newChild := f.remove(f.nodes[branchHash.KeyForMap()], path[len(clone.path)+1:])
-			if !newChild.hasChildren() && !newChild.hasValue() {
-				clone.branches[branchSelector] = nil
-			} else {
-				f.connectChildToParentAndSaveChild(newChild, clone, branchSelector)
-			}
-			clone = f.squash(clone)
-		}
-		return clone
-	}
-	return currentNode
-}
-
-func (f *Forest) add(currentNode *Node, path string, valueHash primitives.Sha256) *Node {
-	newNode := currentNode.clone()
-	if currentNode.path == path { // existing leaf node updated
-		newNode.value = valueHash
-		return newNode
-	}
-
-	if strings.HasPrefix(path, currentNode.path) {
-		if !currentNode.hasValue() && !currentNode.hasChildren() { // this node has no children and no value, replace it
-			newNode = createNode(path, valueHash)
-			return newNode
-		}
-
-		branchSelector := path[len(currentNode.path)]
-		childPath := path[len(currentNode.path)+1:]
-		var newChild *Node
-		if branchHash := currentNode.branches[branchSelector]; branchHash != nil {
-			newChild = f.add(f.nodes[branchHash.KeyForMap()], childPath, valueHash)
-		} else {
-			newChild = createNode(childPath, valueHash)
-		}
-		f.connectChildToParentAndSaveChild(newChild, newNode, branchSelector)
-		return newNode
-	}
-
-	if strings.HasPrefix(currentNode.path, path) { // "insert" a leaf node along the path of currentNode
-		branchSelector := newNode.path[len(path)]
-		newNode.path = newNode.path[len(path)+1:]
-		newParent := createNode(path, valueHash)
-		f.connectChildToParentAndSaveChild(newNode, newParent, branchSelector)
-		return newParent
-	}
-
-	// current node replaced by a new branch node, so that current node is one child and new node is second child
-	i := 0
-	for i = 0; i < len(currentNode.path) && i < len(path) && currentNode.path[i] == path[i]; i++ {
-	}
-	newCommonPath := path[:i]
-	newParent := createNode(newCommonPath, zeroValueHash)
-	newChild := createNode(path[i+1:], valueHash)
-	f.connectChildToParentAndSaveChild(newChild, newParent, path[i])
-
-	newNode.path = newNode.path[i+1:]
-	f.connectChildToParentAndSaveChild(newNode, newParent, currentNode.path[i])
-	return newParent
-}
-
-// appends diffs starting at requested trie node (hash) and returns new trie node hash
-// NOTE: typical use is baseHash is the newest root hash and return value is new root.
-func (f *Forest) Update(baseHash primitives.MerkleSha256, diffs []*protocol.ContractStateDiff) (primitives.MerkleSha256, error) {
-	if _, exists := f.nodes[baseHash.KeyForMap()]; !exists {
-		return nil, errors.Errorf("root node doesn't exist cannot update trie")
-	}
-
-	for _, diff := range diffs {
-		contract := diff.StringContractName()
-		for i := diff.StateDiffsIterator(); i.HasNext(); {
-			record := i.NextStateDiffs()
-			path := contract + record.StringKey()
-			baseHash = f.updateSingleEntry(baseHash, path, hash.CalcSha256([]byte(record.StringValue())))
+	for i := len(f.roots) - 1; i >= 0; i-- {
+		if f.roots[i].hash.Equal(rootHash) {
+			return f.roots[i]
 		}
 	}
-	return baseHash, nil
+
+	return nil
 }
 
-// extract and return a verifiable proof for the value of key in the state snapshot reflected by trieId (corresponding to some block height)
-func (f *Forest) GetProof(rootHash primitives.MerkleSha256, contract string, key string) (Proof, error) {
-	fullPath := contract + key
-	currentNode, exists := f.nodes[rootHash.KeyForMap()]
+func (f *Forest) appendRoot(root *node) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+
+	f.roots = append(f.roots, root)
+}
+
+func (f *Forest) GetProof(rootHash primitives.MerkleSha256, path string) (Proof, error) {
+	current := f.findRoot(rootHash)
+	if current == nil {
+		return nil, errors.Errorf("unknown root")
+	}
+
 	proof := make(Proof, 0, 10)
-	proof = append(proof, currentNode)
+	proof = append(proof, current.serialize())
 
-	for p := fullPath; exists && strings.HasPrefix(p, currentNode.path); {
-		p = p[len(currentNode.path):]
+	for p := path; strings.HasPrefix(p, current.path); {
+		p = p[len(current.path):]
 
 		if p != "" {
-			currentNode, exists = f.nodes[currentNode.branches[p[0]].KeyForMap()]
-			if exists {
-				proof = append(proof, currentNode)
+			if current = current.branches[p[0]]; current != nil {
+				proof = append(proof, current.serialize())
+				p = p[1:]
+			} else {
+				break
 			}
-			p = p[1:]
 		} else {
 			break
 		}
@@ -236,51 +160,171 @@ func (f *Forest) GetProof(rootHash primitives.MerkleSha256, contract string, key
 	return proof, nil
 }
 
-// return true if proof and merkle rootHash validate value for key. false if it confirms value does not match key
-// return an error if the proof is inconsistent internally, or, with key, value or rootHash
-func (f *Forest) Verify(rootHash primitives.MerkleSha256, proof Proof, contract string, key string, value string) (bool, error) {
-	//TODO split the case where we compare against zero value - to simplify determineValueHashByProof
-	valueSha256 := hash.CalcSha256([]byte(value))
-	expectedHash, err := determineValueHashByProof(proof, contract+key, rootHash)
-	if err != nil {
-		return false, err
+func (f *Forest) Verify(rootHash primitives.MerkleSha256, proof Proof, path string, value primitives.Sha256) (bool, error) {
+	currentHash := rootHash
+	emptyMerkleHash := primitives.MerkleSha256{}
+
+	for i, currentNode := range proof {
+		calcHash := currentNode.hash()
+		if !calcHash.Equal(currentHash) { // validate current node against expected hash
+			return false, errors.Errorf("proof hash mismatch at node %d", i)
+		}
+		if path == currentNode.path {
+			return value.Equal(currentNode.value), nil
+		}
+		if len(path) <= len(currentNode.path) {
+			return value.Equal(zeroValueHash), nil
+		}
+		if !strings.HasPrefix(path, currentNode.path) {
+			return value.Equal(zeroValueHash), nil
+		}
+		currentHash = currentNode.branches[path[len(currentNode.path)]]
+		path = path[len(currentNode.path)+1:]
+
+		if emptyMerkleHash.Equal(currentHash) {
+			return value.Equal(zeroValueHash), nil
+		}
 	}
-	return valueSha256.Equal(expectedHash), nil
+
+	return false, errors.Errorf("proof incomplete ")
 }
 
-func determineValueHashByProof(proof Proof, path string, parentHash primitives.MerkleSha256) (primitives.Sha256, error) {
-	if len(proof) == 0 { // proof has ended before a positive conclusion could be reached
-		return nil, errors.Errorf("Proof incomplete")
-	}
+func (f *Forest) Forget(rootHash primitives.MerkleSha256) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	node := proof[0] // each iteration inspects the top (remaining) node in the proof
-
-	if !node.hash().Equal(parentHash) { // validate current node against expected hash
-		return nil, errors.Errorf("Merkle root mismatch or proof may have been tampered with")
-	}
-
-	if path == node.path { // current node consumes the remainder of the key. check hasValue value
-		if node.hasValue() { // key is in trie
-			return node.value, nil
-		} else { // key is not in trie
-			return zeroValueHash, nil
+	newRoots := make([]*node, 0, len(f.roots)-1)
+	for _, root := range f.roots { // naive copy because its a small array and simple code is better
+		if !root.hash.Equal(rootHash) {
+			newRoots = append(newRoots, root)
 		}
-	} else if len(path) <= len(node.path) { // key is not in trie
-		return zeroValueHash, nil
+	}
+	f.roots = newRoots
+}
+
+func (f *Forest) Update(rootMerkle primitives.MerkleSha256, diffs MerkleDiffs) (primitives.MerkleSha256, error) {
+	root := f.findRoot(rootMerkle)
+	if root == nil {
+		return nil, errors.Errorf("must start with valid root")
 	}
 
-	if !strings.HasPrefix(path, node.path) { // key is not in trie
-		return zeroValueHash, nil
+	sandbox := make(dirtyNodes)
+
+	for path, value := range diffs {
+		root = f.travelUpdateAndMark(nil, 0, root, path, value, sandbox)
 	}
 
-	// follow branch: get the hash code of the next expected node for our key
-	nextHash := node.branches[path[len(node.path)]]
-
-	if nextHash == nil { // key is not in trie
-		return zeroValueHash, nil
+	root = f.travelCollapseAndHash(root, sandbox)
+	if root == nil { // special case we got back to empty merkle
+		root = createEmptyNode()
 	}
 
-	// current top node passes validation, proceed to the next node
-	return determineValueHashByProof(proof[1:], path[len(node.path)+1:], nextHash)
+	f.appendRoot(root)
+	return root.hash, nil
+}
 
+func (f *Forest) travelUpdateAndMark(parent *node, arc byte, current *node, path string, valueHash primitives.Sha256, sandbox dirtyNodes) *node {
+	current = f.getOrClone(current, parent, arc, sandbox)
+
+	if current.path == path { // path reached exactly
+		current.value = valueHash
+		return current
+	}
+
+	if strings.HasPrefix(path, current.path) { // current is next part of path
+		if !current.hasValue() && current.isLeaf { // replace it
+			current.path = path
+			current.value = valueHash
+		} else {
+			childArc := path[len(current.path)]
+			childPath := path[len(current.path)+1:]
+			if childNode := current.branches[childArc]; childNode != nil {
+				current.branches[childArc] = f.travelUpdateAndMark(current, childArc, childNode, childPath, valueHash, sandbox)
+			} else if valueHash.Equal(zeroValueHash) {
+				// set to empty value cannot create new children, do nothing
+			} else {
+				newChild := createNode(childPath, valueHash, true)
+				current.branches[childArc] = newChild
+				current.isLeaf = false
+				sandbox.set(current, childArc)
+			}
+		}
+		return current
+	}
+
+	if strings.HasPrefix(current.path, path) { // "insert" a valued node along the path
+		childArc := current.path[len(path)]
+
+		newParent := createNode(path, valueHash, false)
+		newParent.branches[childArc] = current
+		sandbox.set(newParent, childArc)
+
+		current.path = current.path[len(path)+1:]
+		return newParent
+	}
+
+	// new node is a brother of mine so i create a common parent too
+	i := 0
+	for i = 0; i < len(current.path) && i < len(path) && current.path[i] == path[i]; i++ {
+	}
+	newCommonPath := path[:i]
+
+	newParent := createNode(newCommonPath, zeroValueHash, false)
+	newParent.branches[current.path[i]] = current
+	sandbox.set(newParent, current.path[i])
+
+	current.path = current.path[i+1:]
+
+	newChild := createNode(path[i+1:], valueHash, true)
+	newParent.branches[path[i]] = newChild
+	sandbox.set(newParent, path[i])
+
+	return newParent
+}
+
+func (f *Forest) getOrClone(current *node, parent *node, arc byte, sandbox dirtyNodes) *node {
+	var actual *node
+	if exists := sandbox[current]; exists != nil {
+		actual = current
+	} else {
+		actual = current.clone()
+		sandbox.init(actual)
+		sandbox.set(parent, arc)
+	}
+	return actual
+}
+
+func (f *Forest) travelCollapseAndHash(current *node, sandbox dirtyNodes) *node {
+	nChildren := 0
+	aChild := 0
+
+	if !current.isLeaf {
+		for arc := range sandbox[current] {
+			current.branches[arc] = f.travelCollapseAndHash(current.branches[arc], sandbox)
+		}
+
+		// check if i have any children left: count+save last one
+		for arc, child := range current.branches {
+			if child != nil {
+				nChildren++
+				aChild = arc
+			}
+		}
+		current.isLeaf = nChildren == 0
+	}
+
+	// if i have no value ...
+	if !current.hasValue() {
+		if current.isLeaf { // prune empty leaf node
+			return nil
+		} else if nChildren == 1 { // fold up only child
+			child := current.branches[aChild]
+			combinedPath := current.path + string([]byte{byte(aChild)}) + child.path
+			current = child.clone()
+			current.path = combinedPath
+		}
+	}
+
+	current.hash = current.serialize().hash()
+	return current
 }
